Organizing tutorials into working and non-working
The ones that don’t run on my install are going into folders called underDevelopment, otherwise preserving their folder structure.

Some of the cloud ones may work if there is an appropriate cloud interface instaled.  Probably we want some routine that checks whether the cloud is available, and that handles the absence of the cloud interface gracefully.  The same might be true of tutorials that required login access to a flywheel instance for grabbing assetts.