Old ChessSet renamed
The old ChessSet scene is slow to load because the pieces are complex and there are a lot of them.  So we moved the old ChessSet to ChessSetPieces - which lets us manipulate the pieces.  We are replacing ChessSet with a parsed version using PLY files that should be faster to load and run for tests and demos.