Fixed piDat2ISET for ray transfer read.
Not quite done.  There is something wrong for the depth calculation.  But that should be easy to fix when I am not too tired.

The problem is that the depth rendering output is going into the wrong file name, with SimpleScene_coordinates.dat rather than SimpleScene_depth.dat.  No idea why, but surely just something dumb.