Comparing graphics rendering noise with Poisson noise
Looks like we have some more rendering noise that may never go away at the pixels per ray we typically use.

Also fixed an issue with showing the objects when the object is a light.  We should do better on that.

Added a renderingNoise script for the students to use in the Psych 221 class.